backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio>=4.5.2,<5
certifi==2026.1.4
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.124.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pillow==10.4.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
python-telegram-bot==21.6
SQLAlchemy==2.0.45
starlette==0.44.0
typing_extensions==4.13.2
uvicorn==0.33.0
